requirements.txt
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
httpx>=0.11.0
python-slugify>=4.0.0
python-status>=1.0.1
shamir
